old_preprocessing/preprocessing_structure/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "preprocessing"
version = "0.1.0"
description = "Rosbag preprocessing pipeline for bagseek"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcap>=0.0.10",
    "mcap-ros2-support",
    "numpy",
    "pillow",
    "python-dotenv",
    "torch",
    "torchvision",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
]

[project.scripts]
preprocessing = "preprocessing.main:main"

[tool.setuptools]
packages = [
    "preprocessing",
    "preprocessing.collectors",
    "preprocessing.core",
    "preprocessing.postprocessors",
    "preprocessing.processors",
    "preprocessing.processors.mcap",
    "preprocessing.processors.rosbag",
    "preprocessing.utils",
]
package-dir = {"preprocessing" = "."}
